requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
idna==3.3
numpy==1.22.2
pandas==1.4.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
urllib3==1.26.8
